featcHeaderScript file moved in data-loader folder.
