+++ utils/data-loader/fetchHeadScript.ts
